SOF-1552 Refactor architecture of Action- and TimelineObjectFactories
