+++ src/blueprints/tv2/timeline-object-factories/interfaces/tv2-audio-timeline-object-factory.ts
@@ -1,5 +1,5 @@
 import { TimelineObject } from '../../../../model/entities/timeline-object'
-import { Tv2BlueprintConfiguration } from '../tv2-blueprint-configuration'
+import { Tv2BlueprintConfiguration } from '../../value-objects/tv2-blueprint-configuration'
 
 export interface Tv2AudioTimelineObjectFactory {
   createStopAudioBedTimelineObject(duration: number): TimelineObject
