+++ src/blueprints/tv2/timeline-object-factories/interfaces/tv2-graphics-timeline-object-factory.ts
@@ -1,4 +1,4 @@
-import { Tv2BlueprintConfiguration } from '../tv2-blueprint-configuration'
+import { Tv2BlueprintConfiguration } from '../../value-objects/tv2-blueprint-configuration'
 import { TimelineObject } from '../../../../model/entities/timeline-object'
 
 export interface Tv2GraphicsTimelineObjectFactory {
