+++ src/blueprints/tv2/timeline-object-factories/interfaces/tv2-video-mixer-timeline-object-factory.ts
@@ -1,4 +1,4 @@
-import { Tv2DownstreamKeyer } from '../tv2-studio-blueprint-configuration'
+import { Tv2DownstreamKeyer } from '../../value-objects/tv2-studio-blueprint-configuration'
 import { TimelineObject } from '../../../../model/entities/timeline-object'
 
 export interface Tv2VideoMixerTimelineObjectFactory {
