+++ src/blueprints/tv2/timeline-object-factories/tv2-viz-graphics-timeline-object-factory.ts
@@ -1,4 +1,4 @@
-import { Tv2GraphicsTimelineObjectFactory } from '../value-objects/factories/tv2-graphics-timeline-object-factory'
+import { Tv2GraphicsTimelineObjectFactory } from './interfaces/tv2-graphics-timeline-object-factory'
 import { Tv2BlueprintConfiguration } from '../value-objects/tv2-blueprint-configuration'
 import {
   VizMseClearGraphicsTimelineObjectContent,
@@ -109,6 +109,4 @@ export class Tv2VizGraphicsTimelineObjectFactory implements Tv2GraphicsTimelineO
       showName: blueprintConfiguration.showStyle.selectedGraphicsSetup.OvlShowName ?? ''
     }
   }
-
-
 }
