+++ src/blueprints/tv2/tv2-blueprints-facade.ts
@@ -8,10 +8,10 @@ import { Tv2CameraActionFactory } from './factories/tv2-camera-action-factory'
 import { Tv2TransitionActionFactory } from './factories/tv2-transition-action-factory'
 import { Tv2AudioActionFactory } from './factories/tv2-audio-action-factory'
 import { Tv2GraphicsActionFactory } from './factories/tv2-graphics-action-factory'
-import { Tv2VizGraphicsTimelineObjectFactory } from './factories/tv2-viz-graphics-timeline-object-factory'
-import { Tv2SisyfosAudioTimelineObjectFactory } from './factories/tv2-sisyfos-audio-timeline-object-factory'
+import { Tv2VizGraphicsTimelineObjectFactory } from './timeline-object-factories/tv2-viz-graphics-timeline-object-factory'
+import { Tv2SisyfosAudioTimelineObjectFactory } from './timeline-object-factories/tv2-sisyfos-audio-timeline-object-factory'
 import { Tv2VideoMixerConfigurationActionFactory } from './factories/tv2-video-mixer-configuration-action-factory'
-import { Tv2AtemVideoMixerTimelineFactory } from './factories/tv2-atem-video-mixer-timeline-factory'
+import { Tv2AtemVideoMixerTimelineObjectFactory } from './timeline-object-factories/tv2-atem-video-mixer-timeline-object-factory'
 
 export class Tv2BlueprintsFacade {
   public static createBlueprint(): Blueprint {
@@ -23,7 +23,7 @@ export class Tv2BlueprintsFacade {
         new Tv2TransitionActionFactory(),
         new Tv2AudioActionFactory(new Tv2SisyfosAudioTimelineObjectFactory()),
         new Tv2GraphicsActionFactory(new Tv2VizGraphicsTimelineObjectFactory()),
-        new Tv2VideoMixerConfigurationActionFactory(new Tv2AtemVideoMixerTimelineFactory())
+        new Tv2VideoMixerConfigurationActionFactory(new Tv2AtemVideoMixerTimelineObjectFactory())
       )
     )
   }
